+++ src/modbus_client.rs
@@ -28,12 +28,12 @@ use std::io::Error;
 ///     - [`reply_exception()`](struct.Modbus.html#method.reply_exception)
 ///
 pub trait ModbusClient {
-    fn read_bits(&self, address: i32, num: i32, dest: &mut [u8]) -> Result<i32>;
+    fn read_bits(&self, address: u16, num: u16, dest: &mut [u8]) -> Result<i32>;
     fn read_input_bits(&self, address: i32, num: i32, dest: &mut [u8]) -> Result<i32>;
     fn read_registers(&self, address: i32, num: i32, dest: &mut [u16]) -> Result<i32>;
     fn read_input_registers(&self, address: i32, num: i32, dest: &mut [u16]) -> Result<i32>;
     fn report_slave_id(&self, max_dest: usize, dest: &mut [u8]) -> Result<i32>;
-    fn write_bit(&self, address: i32, status: bool) -> Result<()>;
+    fn write_bit(&self, address: u16, status: bool) -> Result<()>;
     fn write_bits(&self, address: i32, num: i32, src: &[u8]) -> Result<i32>;
     fn write_register(&self, address: i32, value: i32) -> Result<()>;
     fn write_registers(&self, address: i32, num: i32, src: &[u16]) -> Result<i32>;
@@ -73,13 +73,13 @@ impl ModbusClient for Modbus {
     ///
     /// assert!(modbus.read_bits(0, 1, &mut dest).is_ok());
     /// ```
-    fn read_bits(&self, address: i32, num: i32, dest: &mut [u8]) -> Result<i32> {
-        if num > dest.len() as i32 {
+    fn read_bits(&self, address: u16, num: u16, dest: &mut [u8]) -> Result<i32> {
+        if num > dest.len() as u16 {
             bail!(ErrorKind::TooManyData("Too many bits requested"));
         }
 
         unsafe {
-            match libmodbus_sys::modbus_read_bits(self.ctx, address as c_int, num, dest.as_mut_ptr()) {
+            match libmodbus_sys::modbus_read_bits(self.ctx, address as c_int, num as c_int, dest.as_mut_ptr()) {
                 -1 => bail!(Error::last_os_error()),
                 len => Ok(len),
             }
@@ -276,7 +276,8 @@ impl ModbusClient for Modbus {
     ///
     /// assert!(modbus.write_bit(address, true).is_ok());
     /// ```
-    fn write_bit(&self, address: i32, status: bool) -> Result<()> {
+    fn write_bit(&self, address: u16, status: bool) -> Result<()> {
+
         unsafe {
             match libmodbus_sys::modbus_write_bit(self.ctx, address as c_int, status as c_int) {
                 -1 => bail!(Error::last_os_error()),
